+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 
 * Fix zeroing pages when incremental savestates feature is disabled
 * Specifying a movie in commandline does not truncate the file anymore
+* Fix a crash when renaming the frame column in input editor
 
 ## [1.3.2] - 2018-11-16
 ### Added